trashDetection/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
filelock==3.12.2
Flask==2.2.5
flatbuffers==24.3.25
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.2
h5py==3.8.0
idna==3.7
importlib-metadata==6.7.0
itsdangerous==2.1.2
jinja2==3.1.4
keras==2.11.0
libclang==18.1.1
Markdown==3.4.4
MarkupSafe==2.1.5
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
Pillow==9.5.0
platformdirs==4.0.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dotenv==0.21.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing-extensions==4.7.1
urllib3==2.0.7
virtualenv==20.26.2
Werkzeug==2.2.3
wrapt==1.16.0
zipp==3.15.0